bugfix: buffer.putInt instead of buffer.putLong.  whoops.
svn path=/autoplot/trunk/QStream/; revision=11613

+++ QStream/src/org/virbo/qstream/LongTransferType.java
@@ -18,7 +18,7 @@ public class LongTransferType extends TransferType {
 
     @Override
     public void write(double d, ByteBuffer buffer) {
-        buffer.putLong((int)d);
+        buffer.putLong((long)d);
     }
 
     @Override
